# app/documentation/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/new_job":{"get":{"summary":"Read Root","operationId":"read_root_new_job_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/item":{"get":{"summary":"Get Item","operationId":"get_item_item_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Item"},"type":"array","title":"Response Get Item Item Get"}}}}}}},"/items/search/":{"get":{"tags":["Items"],"summary":"Search Items By Ocr","description":"Busca items por el campo 'ocr'.\n\n:param ocr: El valor del campo 'ocr' a buscar.\n:return: Lista de items que coinciden con el 'ocr' proporcionado.","operationId":"search_items_by_ocr_items_search__get","parameters":[{"name":"ocr","in":"query","required":true,"schema":{"type":"string","title":"Ocr"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemRead"},"title":"Response Search Items By Ocr Items Search  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ocr/":{"post":{"tags":["OCR"],"summary":"Extract Text","description":"Extrae texto de una imagen usando Azure OCR.\n\n:param image: Archivo de imagen cargado por el usuario.\n:return: Lista de textos reconocidos en la imagen.","operationId":"extract_text_ocr__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_extract_text_ocr__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/all_jobs":{"get":{"tags":["Jobs"],"summary":"Get Item","description":"Devuelve una lista de todos los Jobs_codes existentes.","operationId":"get_item_jobs_all_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Job"},"type":"array","title":"Response Get Item Jobs All Jobs Get"}}}}}}},"/jobs/job_available/{job_code}":{"get":{"tags":["Jobs"],"summary":"Check Job Available","description":"Verifica si un job_code está disponible.\nRetorna True si el job_code no existe en la base de datos,\nFalse si ya está en uso.","operationId":"check_job_available_jobs_job_available__job_code__get","parameters":[{"name":"job_code","in":"path","required":true,"schema":{"type":"string","title":"Job Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"boolean","title":"Response Check Job Available Jobs Job Available  Job Code  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/job_exists/{job_code}":{"get":{"tags":["Jobs"],"summary":"Check Job Exists","description":"Verifica si un job_code existe.\nRetorna True si el job_code existe en la base de datos,\nFalse si no existe.","operationId":"check_job_exists_jobs_job_exists__job_code__get","parameters":[{"name":"job_code","in":"path","required":true,"schema":{"type":"string","title":"Job Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"boolean","title":"Response Check Job Exists Jobs Job Exists  Job Code  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/upload_csv/":{"post":{"tags":["Jobs"],"summary":"Upload Items Csv","operationId":"upload_items_csv_jobs_upload_csv__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_items_csv_jobs_upload_csv__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/object/data/{job_code}":{"get":{"tags":["Object"],"summary":"Get Dummy Data","description":"Devuelve la lista de datos dummy independientemente del job_code.","operationId":"get_dummy_data_object_data__job_code__get","parameters":[{"name":"job_code","in":"path","required":true,"schema":{"type":"string","title":"Job Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/test_job":{"get":{"tags":["Jobs"],"summary":"Get Dummy Data","description":"Devuelve la lista de datos dummy.","operationId":"get_dummy_data_jobs_test_job_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/objects/job_items/{job_code}":{"get":{"tags":["Object"],"summary":"Check Job Available","description":"Dado un job_code obtiene los objetos asociados, mostrando sus OCR y la cantidad de estos.","operationId":"check_job_available_objects_job_items__job_code__get","parameters":[{"name":"job_code","in":"path","required":true,"schema":{"type":"string","title":"Job Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ObjectResponse"},"title":"Response Check Job Available Objects Job Items  Job Code  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_extract_text_ocr__post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_extract_text_ocr__post"},"Body_upload_items_csv_jobs_upload_csv__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_items_csv_jobs_upload_csv__post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Item":{"properties":{"item_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Item Id"},"object_id":{"type":"integer","title":"Object Id"},"current_stage":{"type":"integer","title":"Current Stage"},"rework":{"type":"integer","title":"Rework","default":0},"scrap":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Scrap","default":0}},"type":"object","required":["object_id","current_stage"],"title":"Item"},"ItemRead":{"properties":{"object_id":{"type":"integer","title":"Object Id"},"current_stage":{"type":"integer","title":"Current Stage"},"ocr":{"type":"string","title":"Ocr"},"rework":{"type":"integer","title":"Rework"},"scrap":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Scrap","default":0},"item_id":{"type":"integer","title":"Item Id"}},"type":"object","required":["object_id","current_stage","ocr","rework","item_id"],"title":"ItemRead"},"Job":{"properties":{"job_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Job Id"},"job_code":{"type":"string","title":"Job Code"},"client_id":{"type":"integer","title":"Client Id"},"status":{"type":"boolean","title":"Status","default":false},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["job_code","client_id"],"title":"Job"},"ObjectResponse":{"properties":{"ocr":{"type":"string","title":"Ocr"},"cantidad":{"type":"string","title":"Cantidad"}},"type":"object","required":["ocr","cantidad"],"title":"ObjectResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}